broker: Improvements for manipulating concentrations
Let get/set sequence work on free concentration of the given sequence.

Add many checks for indices, positive quantities.

+++ src/mirbooking-default-score-table.c
@@ -153,7 +153,7 @@ static gfloat
 binding_energy (gfloat *G, gsize n)
 {
     guint i;
-    gdouble Gtot = 0, Z = 0;
+    gfloat Gtot = 0, Z = 0;
     for (i = 0; i < n; i++)
     {
         if (isfinite (G[i]))
